pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eval-kit"
version = "0.1.0"
description = "Layered filter-chain evaluation toolkit for LLM outputs."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "BSL-1.1"}
authors = [
    {name = "Alex Sawinski", email = "sawinskia23@gmail.com"},
]
keywords = ["llm", "evaluation", "nlp", "fine-tuning", "quality"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]
rubrics = [
    "pyyaml>=6.0",
]
all = [
    "pyyaml>=6.0",
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.scripts]
eval-kit = "eval_kit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
